Applied HARMONY-3321 [drlvm] warnings when building encoder on em64t
git-svn-id: https://svn.apache.org/repos/asf/harmony/enhanced/drlvm/trunk@515646 13f79535-47bb-0310-9956-ffa450edef68

+++ vm/em/src/EdgeProfileCollector.cpp
@@ -61,7 +61,7 @@ uint32 edge_profiler_get_num_counters(Method_Profile_Handle mph)
 {
     MethodProfile* mp = (MethodProfile*)mph;
     assert(mp->pc->type == EM_PCTYPE_EDGE);
-    return ((EdgeMethodProfile*)mp)->counters.size();
+    return (uint32)((EdgeMethodProfile*)mp)->counters.size();
 }
